spacerunner/requirements.txt
# GRPO training stack (roadmap §M6).
# Heavy ML deps; install in a separate venv from the env server.
# Pinned per the roadmap table; unsloth requires CUDA + recent PyTorch.
trl>=0.11
transformers>=4.45
accelerate>=0.34
peft>=0.12
bitsandbytes>=0.43
unsloth>=2024.10
wandb>=0.18
pyyaml>=6.0
# Already required by the env client, repeated here so `pip install -r requirements-train.txt`
# is self-sufficient on a fresh training node.
httpx>=0.27
pydantic>=2.7
openenv-core>=0.2.3
fastapi>=0.110
uvicorn>=0.30
huggingface_hub>=0.24
tensorboard>=2.16
